+++ ccxt.php
@@ -211,6 +211,7 @@ class Exchange {
         'quoine',
         'southxchange',
         'surbitcoin',
+        'tidex',
         'therock',
         'urdubit',
         'vaultoro',